+++ python/ray/train/context.py
@@ -3,11 +3,11 @@ from typing import TYPE_CHECKING, Any, Dict, Optional
 
 from ray.train._internal import session
 from ray.train._internal.storage import StorageContext
-from ray.train.utils import _copy_doc, _log_deprecation_warning
-from ray.train.v2._internal.migration_utils import (
+from ray.train.constants import (
     _v2_migration_warnings_enabled,
     V2_MIGRATION_GUIDE_MESSAGE,
 )
+from ray.train.utils import _copy_doc, _log_deprecation_warning
 from ray.util.annotations import Deprecated, DeveloperAPI, PublicAPI
 
 if TYPE_CHECKING:
